Update create.sql
So, like the newb I am at this, I committed to the master and not your working branch (which I didn't find until I was too far in).  Feel free to alter or haggle over the differences and we can go with the newest commit tomorrow afternoon (I know how to find that now).

I keep going back and forth between having the check on the position name or not thinking it would help avoid possible duplicates from entry but also limits ability to add other jobs in the future (such as adapting for officers).